+++ src/NetAppFiles/NetAppFiles/ActiveDirectory/NewNetAppFilesActiveDirectory.cs
@@ -131,6 +131,26 @@ namespace Microsoft.Azure.Commands.NetAppFiles.BackupPolicy
         [ValidateNotNullOrEmpty]
         public string[] SecurityOperator { get; set; }
 
+
+        [Parameter(
+            Mandatory = false,
+            HelpMessage = "When AES is enabled, set if AES encryption will be enabled for SMB communication.")]
+        [ValidateNotNullOrEmpty]
+        public SwitchParameter AesEncryption { get; set; }
+
+        [Parameter(
+            Mandatory = false,
+            HelpMessage = "When LDAP over SSL/TLS is enabled, Specifies whether or not the LDAP traffic needs to be signed.")]
+        [ValidateNotNullOrEmpty]
+        public SwitchParameter LdapSigning { get; set; }
+
+        [Parameter(
+            Mandatory = false,
+            HelpMessage = "When LDAP over SSL/TLS is enabled, specifies whether or not the LDAP traffic needs to be secured via TLS.")]
+        [ValidateNotNullOrEmpty]
+        public SwitchParameter LdapOverTLS { get; set; }
+        
+
         [Parameter(
             ParameterSetName = ParentObjectParameterSet,
             Mandatory = true,
@@ -169,7 +189,10 @@ namespace Microsoft.Azure.Commands.NetAppFiles.BackupPolicy
                     BackupOperators = BackupOperator,
                     KdcIP = KdcIP,
                     ServerRootCACertificate = ServerRootCACertificate,
-                    SecurityOperators = SecurityOperator
+                    SecurityOperators = SecurityOperator,
+                    AesEncryption = AesEncryption,
+                    LdapSigning = LdapSigning,
+                    LdapOverTLS = LdapOverTLS
                 };
                 if (anfAccount.ActiveDirectories == null)
                 {