+++ src/NetAppFiles/NetAppFiles/BackupPolicy/SetNetAppFilesBackupPolicy.cs
@@ -23,17 +23,16 @@ using System.Globalization;
 using Microsoft.Azure.Commands.NetAppFiles.Helpers;
 using Microsoft.Azure.Management.Monitor.Version2018_09_01.Models;
 using System.Collections.Generic;
-using System;
 
 namespace Microsoft.Azure.Commands.NetAppFiles.BackupPolicy
 {
     [Cmdlet(
-        "New",
+        "Set",
         ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "NetAppFilesBackupPolicy",
         SupportsShouldProcess = true,
         DefaultParameterSetName = FieldsParameterSet), OutputType(typeof(PSNetAppFilesBackupPolicy))]
-    [Alias("New-AnfBackupPolicy")]
-    public class NewAzureRmNetAppFilesBackupPolicy : AzureNetAppFilesCmdletBase
+    [Alias("Set-AnfBackupPolicy")]
+    public class SetAzureRmNetAppFilesBackupPolicy : AzureNetAppFilesCmdletBase
     {
         [Parameter(
             Mandatory = true,
@@ -136,21 +135,6 @@ namespace Microsoft.Azure.Commands.NetAppFiles.BackupPolicy
                 {
                     tagPairs.Add(key, Tag[key].ToString());
                 }
-            }
-
-            Management.NetApp.Models.BackupPolicy existingBackupPolicy = null;
-
-            try
-            {
-                existingBackupPolicy = AzureNetAppFilesManagementClient.BackupPolicies.Get(ResourceGroupName, AccountName, Name);
-            }
-            catch
-            {
-                existingBackupPolicy = null;
-            }
-            if (existingBackupPolicy != null)
-            {
-                throw new Exception(string.Format("A Backup Policy with name '{0}' in resource group '{1}' already exists. Please use Set/Update-AzNetAppFilesBackupPolicy to update an existing Backup Policy.", this.Name, this.ResourceGroupName));
             }
 
             var backupPolicyBody = new Management.NetApp.Models.BackupPolicy()