+++ src/NetAppFiles/NetAppFiles/Volume/NewNetAppFilesVolume.cs
@@ -21,6 +21,7 @@ using Microsoft.Azure.Commands.NetAppFiles.Models;
 using Microsoft.Azure.Management.NetApp;
 using System.Collections.Generic;
 using Microsoft.WindowsAzure.Commands.Common.CustomAttributes;
+using System;
 
 namespace Microsoft.Azure.Commands.NetAppFiles.Volume
 {